+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+* Basic documentation added.
+* Basic example project added.
+* Unnecessary notification listeners removed.
+* State updater is based only in the centralized call of setState((){}) for the builders linked to a Fx variable.
+
 ## 0.0.2
 * RouteTransition delimited to animated and none (linear, curve and fade removed). Support for curves added in the transition animation.
 * FxRouter: Basic navigations actions added (goTo, goToAndReplace, back, backUntil).